+++ packages/kit/src/core/postbuild/analyse.js
@@ -122,15 +122,7 @@ async function analyse({ manifest_path, env }) {
 				validate_common_exports(page.universal, page.universal_id);
 			}
 
-			const should_prerender = get_option(nodes, 'prerender');
-			prerender =
-				should_prerender === true ||
-				// Try prerendering if ssr is false and no server needed. Set it to 'auto' so that
-				// the route is not removed from the manifest, there could be a server load function.
-				// People can opt out of this behavior by explicitly setting prerender to false
-				(should_prerender !== false && get_option(nodes, 'ssr') === false && !page?.server?.actions
-					? 'auto'
-					: should_prerender ?? false);
+			prerender = get_option(nodes, 'prerender') ?? false;
 
 			config = get_config(nodes);
 		}