+++ packages/kit/test/apps/basics/src/routes/actions/enhance/+page.svelte
@@ -14,12 +14,14 @@
 <pre class="formdata2">{JSON.stringify($page.form)}</pre>
 
 <form method="post" action="?/login" use:enhance>
-	<input type=hidden name="action" value="DOM clobbering" />
-	<input type=hidden name="reset" value="DOM clobbering" />
+	<input type="hidden" name="action" value="DOM clobbering" />
+	<input type="hidden" name="reset" value="DOM clobbering" />
 	<input name="username" type="text" />
 	<button class="form1">Submit</button>
 	<button class="form1-register" formAction="?/register">Submit</button>
-	<button class="form1-submitter" formAction="?/submitter" name="submitter" value="foo">Submit</button>
+	<button class="form1-submitter" formAction="?/submitter" name="submitter" value="foo">
+		Submit
+	</button>
 	<button class="form1-error" formAction="?/error">Submit</button>
 </form>
 
@@ -34,4 +36,9 @@
 	}}
 >
 	<button class="form2">Submit</button>
+</form>
+
+<form method="post" action="?/echo" use:enhance>
+	<input name="message" type="text" value={form?.message ?? ''} />
+	<button class="form3">Submit</button>
 </form>