+++ packages/kit/test/prerendering/ssr-false/src/routes/prerenderable-2/+page.svelte
@@ -1,1 +1,0 @@
-<p>prerenderable shell</p>