+++ packages/kit/test/prerendering/ssr-false/src/routes/prerenderable/+page.js
@@ -1,3 +1,0 @@
-export function load() {
-	throw new Error('I should not be called during prerendering');
-}