+++ sites/kit.svelte.dev/src/lib/docs/server/index.js
@@ -8,7 +8,7 @@ import 'prismjs/components/prism-typescript.js';
 import 'prism-svelte';
 import { escape, extract_frontmatter, transform } from './markdown.js';
 import { modules } from './type-info.js';
-import { render, replace_placeholders } from './render.js';
+import { replace_placeholders } from './render.js';
 import { parse_route_id } from '../../../../../../packages/kit/src/utils/routing.js';
 import ts from 'typescript';
 import MagicString from 'magic-string';
@@ -162,6 +162,10 @@ export async function read_file(file) {
 					// from e.g. ambient.d.ts
 					if (file.endsWith('30-modules.md')) {
 						injected.push('// @errors: 7006 7031');
+					}
+
+					if (file.endsWith('10-configuration.md')) {
+						injected.push('// @errors: 2307');
 					}
 
 					// another special case
@@ -480,14 +484,18 @@ function convert_to_ts(js_code, indent = '', offset = '') {
 						const [name, generics] = get_type_info(tag);
 
 						if (ts.isFunctionDeclaration(node)) {
-							const is_export = node.modifiers?.some(
-								(modifier) => modifier.kind === ts.SyntaxKind.ExportKeyword
-							)
-								? 'export '
-								: '';
-							const is_async = node.modifiers?.some(
-								(modifier) => modifier.kind === ts.SyntaxKind.AsyncKeyword
-							);
+							const is_export =
+								ts.canHaveModifiers(node) &&
+								ts
+									.getModifiers(node)
+									?.some((modifier) => modifier.kind === ts.SyntaxKind.ExportKeyword)
+									? 'export '
+									: '';
+							const is_async =
+								ts.canHaveModifiers(node) &&
+								ts
+									.getModifiers(node)
+									?.some((modifier) => modifier.kind === ts.SyntaxKind.AsyncKeyword);
 							code.overwrite(
 								node.getStart(),
 								node.name.getEnd(),
@@ -580,9 +588,11 @@ function convert_to_ts(js_code, indent = '', offset = '') {
 				imports.set(from, new Set([name]));
 			}
 			if (generics !== undefined) {
-				return [name, generics
-					.replaceAll("*", "") // get rid of JSDoc asterisks
-					.replace("  }>", "}>") // unindent closing brace
+				return [
+					name,
+					generics
+						.replaceAll('*', '') // get rid of JSDoc asterisks
+						.replace('  }>', '}>') // unindent closing brace
 				];
 			}
 		}